refactor(controller): 消除 koduck-backend Controller 层冗余
- 删除 AShareController.java（功能被 MarketController + KlineController 覆盖）
- 删除 ProfileController.java（空壳实现，功能被 UserController + SettingsController 替代）
- 提取 LatestPriceResponse 到 dto/market/ 包作为共享 DTO
- 更新 KlineController 引用新的共享 LatestPriceResponse

Closes #384
